+++ src/assets/scripts/client/aircraft/FlightManagementSystem/Fms.js
@@ -456,7 +456,11 @@ export default class Fms {
 
     activateHoldForWaypointName(waypointName, holdParameters) {
         if (!this._routeModel.hasWaypointName(waypointName)) {
-            return [false, `unable to hold at ${waypointName}; it is not on our route!`];
+            // force lower-case in verbal readback to get speech synthesis to pronounce the fix instead of speling it
+            return [false, {
+                log: `unable to hold at ${waypointName}; it is not on our route!`,
+                say: `unable to hold at ${waypointName.toLowerCase()}; it is not on our route!`
+            }];
         }
 
         this._routeModel.activateHoldForWaypointName(waypointName, holdParameters);
@@ -482,6 +486,20 @@ export default class Fms {
         }
 
         return this._routeModel.absorbRouteModel(nextRouteModel);
+    }
+
+    /**
+     * Return the waypoint that matches the provided name
+     *
+     * @for Fms
+     * @method findWaypoiont
+     * @param waypointName {string} name of the waypoint
+     * @return {WaypointModel}
+     */
+    findWaypoint(waypointName) {
+        const name = waypointName.toUpperCase();
+
+        return _find(this.waypoints, (waypoint) => waypoint.name === name);
     }
 
     /**
@@ -1076,12 +1094,12 @@ export default class Fms {
 
         if (!this._routeModel.isRunwayModelValidForStar(nextRunwayModel)) {
             const readback = {};
-            readback.log = `unable, according to our charts, Runway ${nextRunwayModel.name} is ` +
-                            `not valid for the ${this._routeModel.getStarIcao()} arrival, expecting ` +
-                            `Runway ${currentArrivalRunway.name} instead`;
-            readback.say = `unable, according to our charts, Runway ${nextRunwayModel.getRadioName()} ` +
-                            `is not valid for the ${this._routeModel.getStarName()} arrival, expecting ` +
-                            `Runway ${currentArrivalRunway.getRadioName()} instead`;
+            readback.log = `unable, according to our charts, Runway ${nextRunwayModel.name} is `
+                + `not valid for the ${this._routeModel.getStarIcao()} arrival, expecting `
+                + `Runway ${currentArrivalRunway.name} instead`;
+            readback.say = `unable, according to our charts, Runway ${nextRunwayModel.getRadioName()} `
+                + `is not valid for the ${this._routeModel.getStarName()} arrival, expecting `
+                + `Runway ${currentArrivalRunway.getRadioName()} instead`;
 
             return [false, readback];
         }
@@ -1145,8 +1163,8 @@ export default class Fms {
      */
     _verifyRouteContainsMultipleWaypoints() {
         if (this.waypoints.length < 2) {
-            throw new TypeError('Expected flight plan route to have at least two ' +
-                `waypoints, but only found ${this.waypoints.length} waypoints`);
+            throw new TypeError('Expected flight plan route to have at least two '
+                + `waypoints, but only found ${this.waypoints.length} waypoints`);
         }
     }
 }