+++ src/assets/scripts/client/airport/AirportModel.js
@@ -1,6 +1,7 @@
 /* eslint-disable max-len */
 import _ceil from 'lodash/ceil';
 import _chunk from 'lodash/chunk';
+import _clamp from 'lodash/clamp';
 import _forEach from 'lodash/forEach';
 import _get from 'lodash/get';
 import _head from 'lodash/head';
@@ -223,34 +224,34 @@ export default class AirportModel {
         };
 
         /**
+         * @for AirportModel
          * @property ctr_radius
          * @type {nunmber}
-         * @default DEFAULT_CTR_RADIUS_KM
          */
-        this.ctr_radius = DEFAULT_CTR_RADIUS_KM;
+        this.ctr_radius = null;
 
         /**
+         * @for AirportModel
          * @property ctr_ceiling
          * @type {nunmber}
-         * @default DEFAULT_CTR_CEILING_FT
          */
-        this.ctr_ceiling = DEFAULT_CTR_CEILING_FT;
+        this.ctr_ceiling = null;
 
         /**
+         * @for AirportModel
          * @property initial_alt
          * @type {nunmber}
-         * @default DEFAULT_INITIAL_ALTITUDE_FT
          */
-        this.initial_alt = DEFAULT_INITIAL_ALTITUDE_FT;
+        this.initial_alt = null;
 
         /**
+         * @for AirportModel
          * @property rangeRings
          * @type {object}
-         * @default DEFAULT_RANGE_RINGS
          */
-        this.rangeRings = DEFAULT_RANGE_RINGS;
+        this.rangeRings = null;
 
-        this.parse(options);
+        this.init(options);
     }
 
     /**
@@ -343,10 +344,10 @@ export default class AirportModel {
 
     /**
      * @for AirportModel
-     * @method parse
+     * @method init
      * @param data {object}
      */
-    parse(data) {
+    init(data) {
         this.name = _get(data, 'name', this.name);
         this.icao = _get(data, 'icao', this.icao).toLowerCase();
         this.level = _get(data, 'level', this.level);
@@ -365,7 +366,7 @@ export default class AirportModel {
         this.ctr_radius = _get(data, 'ctr_radius', DEFAULT_CTR_RADIUS_KM);
         this.ctr_ceiling = _get(data, 'ctr_ceiling', DEFAULT_CTR_CEILING_FT);
         this.initial_alt = _get(data, 'initial_alt', DEFAULT_INITIAL_ALTITUDE_FT);
-        this.rangeRings = _get(data, 'rangeRings');
+        this.rangeRings = _get(data, 'rangeRings', DEFAULT_RANGE_RINGS);
         this._runwayCollection = new RunwayCollection(data.runways, this._positionModel);
 
         this.loadTerrain();
@@ -415,16 +416,15 @@ export default class AirportModel {
         this.perimeter = _head(this.airspace);
         this.ctr_radius = Math.max(
             ..._map(this.perimeter.poly, (vertexPosition) => vlen(
-                    vsub(
-                        vertexPosition.relativePosition,
-                        DynamicPositionModel.calculateRelativePosition(
-                            this.rangeRings.center,
-                            this._positionModel,
-                            this.magneticNorth
-                        )
+                vsub(
+                    vertexPosition.relativePosition,
+                    DynamicPositionModel.calculateRelativePosition(
+                        this.rangeRings.center,
+                        this._positionModel,
+                        this.magneticNorth
                     )
                 )
-            )
+            ))
         );
     }
 
@@ -505,6 +505,18 @@ export default class AirportModel {
 
             this.restricted_areas.push(obj);
         });
+    }
+
+    /**
+     * Return an altitude clamped within the min/max assignable altitudes for this airport
+     *
+     * @for AirportModel
+     * @method clampWithinAssignableAltitudes
+     * @param altitude
+     * @return {number}
+     */
+    clampWithinAssignableAltitudes(altitude) {
+        return _clamp(altitude, this.minAssignableAltitude, this.maxAssignableAltitude);
     }
 
     /**
@@ -639,8 +651,8 @@ export default class AirportModel {
             return;
         }
 
-        console.warn('Did not expect a query for runway that applies to aircraft of category ' +
-            `'${category}'! Returning the arrival runway (${this.arrivalRunwayModel.name})`);
+        console.warn('Did not expect a query for runway that applies to aircraft of category '
+            + `'${category}'! Returning the arrival runway (${this.arrivalRunwayModel.name})`);
 
         return this.arrivalRunwayModel;
     }
@@ -766,17 +778,14 @@ export default class AirportModel {
         zlsa.atc.loadAsset({
             url: `assets/airports/terrain/${this.icao.toLowerCase()}.geojson`,
             immediate: true
-        })
-        // TODO: change to onSuccess and onError handler abstractions
-        .done((data) => {
+        }).done((data) => { // TODO: change to onSuccess and onError handler abstractions
             try {
                 // eslint-disable-next-line no-undef
                 this.parseTerrain(data);
             } catch (e) {
                 throw new Error(e.message);
             }
-        })
-        .fail((jqXHR, textStatus, errorThrown) => {
+        }).fail((jqXHR, textStatus, errorThrown) => {
             console.error(`Unable to load airport/terrain/${this.icao}: ${textStatus}`);
 
             this.loading = false;
@@ -809,9 +818,8 @@ export default class AirportModel {
         zlsa.atc.loadAsset({
             url: `assets/airports/${this.icao.toLowerCase()}.json`,
             immediate: true
-        })
-        .done((response) => this.onLoadAirportSuccess(response))
-        .fail((...args) => this.onLoadAirportError(...args));
+        }).done((response) => this.onLoadAirportSuccess(response))
+            .fail((...args) => this.onLoadAirportError(...args));
     }
 
     /**
@@ -824,7 +832,7 @@ export default class AirportModel {
         this.loading = false;
         this.loaded = true;
 
-        this.parse(response);
+        this.init(response);
         this.eventBus.trigger(EVENT.AIRPORT_CHANGE, this.data);
         this.set();
     };
@@ -859,7 +867,7 @@ export default class AirportModel {
         this.loading = false;
         this.loaded = true;
 
-        this.parse(response);
+        this.init(response);
         this.set();
     }
 }