+++ src/assets/scripts/client/game/GameController.js
@@ -566,10 +566,18 @@ class GameController {
      */
     _onWindowBlur(event) {
         this.game.focused = false;
+
         // resetting back to 1 here so when focus returns, we can reliably reset
         // `#game.delta` to 0 to prevent jumpiness
         TimeKeeper.updateSimulationRate(1);
         TimeKeeper.setPause(true);
+
+        // update visual state of the timewarp control button for consistency
+        const $fastForwards = $(SELECTORS.DOM_SELECTORS.FAST_FORWARDS);
+
+        $fastForwards.removeClass(SELECTORS.CLASSNAMES.SPEED_2);
+        $fastForwards.removeClass(SELECTORS.CLASSNAMES.SPEED_5);
+        $fastForwards.prop('title', 'Set time warp to 2');
     }
 
     /**
@@ -580,6 +588,11 @@ class GameController {
      */
     _onWindowFocus(event) {
         this.game.focused = true;
+
+        // if was already manually paused when lost focus, respect that
+        if ($('html').hasClass(SELECTORS.CLASSNAMES.PAUSED)) {
+            return;
+        }
 
         TimeKeeper.setPause(false);
     }