+++ test/aircraft/Pilot/Pilot.spec.js
@@ -63,6 +63,17 @@ function buildPilotWithComplexRoute() {
     return pilot;
 }
 
+let sandbox;
+/* eslint-disable no-unused-vars, no-undef */
+ava.beforeEach(() => {
+    sandbox = sinon.sandbox.create();
+});
+
+ava.afterEach(() => {
+    sandbox.restore();
+});
+/* eslint-enable no-unused-vars, no-undef */
+
 ava('throws when instantiated without parameters', (t) => {
     t.throws(() => new Pilot());
     t.throws(() => new Pilot({}));
@@ -78,6 +89,15 @@ ava('throws when instantiated without parameters', (t) => {
 
 ava('does not throw when passed valid parameters', (t) => {
     t.notThrows(() => createPilotFixture());
+});
+
+ava('.reset() properly resets the instance properties to their null state', (t) => {
+    const pilotModel = createPilotFixture().reset();
+
+    t.true(pilotModel._fms === null);
+    t.true(pilotModel._mcp === null);
+    t.true(pilotModel.hasApproachClearance === false);
+    t.true(pilotModel.hasDepartureClearance === false);
 });
 
 ava('.shouldExpediteAltitudeChange() sets #shouldExpediteAltitudeChange to true and responds with a success message', (t) => {
@@ -295,14 +315,14 @@ ava('.applyPartialRouteAmendment() grants departure clearance when the route ame
     t.true(departureClearance);
 });
 
-ava('.applyPartialRouteAmendment() calls .exitHold()', (t) => {
+ava('.applyPartialRouteAmendment() calls .cancelHoldingPattern()', (t) => {
     const pilot = buildPilotWithComplexRoute();
-    const exitHoldSpy = sinon.spy(pilot, 'exitHold');
+    const cancelHoldingPatternSpy = sinon.spy(pilot, 'cancelHoldingPattern');
 
     pilot.initiateHoldingPattern('MISEN', holdParametersMock);
     pilot.applyPartialRouteAmendment(amendRouteString);
 
-    t.true(exitHoldSpy.calledWithExactly());
+    t.true(cancelHoldingPatternSpy.calledWithExactly());
 });
 
 ava('.cancelApproachClearance() returns early if #hasApproachClearance is false', (t) => {
@@ -364,6 +384,84 @@ ava('.cancelApproachClearance() sets #hasApproachClearance to false', (t) => {
     t.false(aircraftModel.pilot.hasApproachClearance);
 });
 
+ava('.cancelHoldingPattern() returns error response when the aircraft does not have any hold waypoints', (t) => {
+    const pilot = createPilotFixture();
+    const expectedResult = [false, 'that must be for somebody else, we weren\'t given any holding instructions'];
+    const result = pilot.cancelHoldingPattern();
+
+    t.deepEqual(result, expectedResult);
+});
+
+ava('.cancelHoldingPattern() returns error response when the aircraft has holding, but the specified fix is not on the route', (t) => {
+    const pilot = createPilotFixture();
+
+    pilot.initiateHoldingPattern('KEPEC', holdParametersMock);
+
+    const expectedResult = [false, {
+        log: 'that must be for somebody else, we weren\'t given holding over ABCDE',
+        say: 'that must be for somebody else, we weren\'t given holding over abcde'
+    }];
+    const result = pilot.cancelHoldingPattern('ABCDE');
+
+    t.deepEqual(result, expectedResult);
+});
+
+ava('.cancelHoldingPattern() returns error response when the aircraft has holding, but not at the specified fix', (t) => {
+    const pilot = createPilotFixture();
+
+    pilot.initiateHoldingPattern('KEPEC', holdParametersMock);
+
+    const expectedResult = [false, {
+        log: 'that must be for somebody else, we weren\'t given holding over SUNST',
+        say: 'that must be for somebody else, we weren\'t given holding over sunst'
+    }];
+    const result = pilot.cancelHoldingPattern('SUNST');
+
+    t.deepEqual(result, expectedResult);
+});
+
+ava('.cancelHoldingPattern() calls WaypointModel.deactivateHold() when no hold fix is specified', (t) => {
+    const pilot = createPilotFixture();
+    const currentWaypointModel = pilot._fms.currentWaypoint;
+    const holdWaypointName = 'KEPEC';
+    const holdWaypointModel = pilot._fms.findWaypoint(holdWaypointName);
+    const currentWaypointDeactivateHoldStub = sinon.stub(currentWaypointModel, 'deactivateHold');
+    const holdWaypointDeactivateHoldStub = sinon.stub(holdWaypointModel, 'deactivateHold');
+
+    pilot.initiateHoldingPattern('KEPEC', holdParametersMock);
+
+    const expectedResult = [true, {
+        log: 'roger, we\'ll cancel the hold at KEPEC',
+        say: 'roger, we\'ll cancel the hold at kepec'
+    }];
+    const result = pilot.cancelHoldingPattern();
+
+    t.deepEqual(result, expectedResult);
+    t.true(currentWaypointDeactivateHoldStub.notCalled);
+    t.true(holdWaypointDeactivateHoldStub.calledWithExactly());
+});
+
+ava('.cancelHoldingPattern() calls WaypointModel.deactivateHold() when the fix is specified by the user', (t) => {
+    const pilot = createPilotFixture();
+    const currentWaypointModel = pilot._fms.currentWaypoint;
+    const holdWaypointName = 'KEPEC';
+    const holdWaypointModel = pilot._fms.findWaypoint(holdWaypointName);
+    const currentWaypointDeactivateHoldStub = sinon.stub(currentWaypointModel, 'deactivateHold');
+    const holdWaypointDeactivateHoldStub = sinon.stub(holdWaypointModel, 'deactivateHold');
+
+    pilot.initiateHoldingPattern('KEPEC', holdParametersMock);
+
+    const expectedResult = [true, {
+        log: 'roger, we\'ll cancel the hold at KEPEC',
+        say: 'roger, we\'ll cancel the hold at kepec'
+    }];
+    const result = pilot.cancelHoldingPattern('KEPEC');
+
+    t.deepEqual(result, expectedResult);
+    t.true(currentWaypointDeactivateHoldStub.notCalled);
+    t.true(holdWaypointDeactivateHoldStub.calledWithExactly());
+});
+
 ava('.clearedAsFiled() grants pilot departure clearance and returns the correct response strings', (t) => {
     const aircraftModel = new AircraftModel(DEPARTURE_AIRCRAFT_INIT_PROPS_MOCK, createNavigationLibraryFixture());
     const result = aircraftModel.pilot.clearedAsFiled();
@@ -377,13 +475,7 @@ ava('.clearedAsFiled() grants pilot departure clearance and returns the correct 
 });
 
 ava('.climbViaSID() returns error response if #flightPlanAltitude has not been set', (t) => {
-    const expectedResult = [
-        false,
-        {
-            log: 'unable to climb via SID, no altitude assigned',
-            say: 'unable to climb via SID, no altitude assigned'
-        }
-    ];
+    const expectedResult = [false, 'unable, no altitude assigned'];
     const aircraftModel = new AircraftModel(DEPARTURE_AIRCRAFT_INIT_PROPS_MOCK);
     aircraftModel.altitude = 0;
     const pilot = new Pilot(createFmsDepartureFixture(), createModeControllerFixture(), createNavigationLibraryFixture());
@@ -397,19 +489,22 @@ ava('.climbViaSID() returns error response if #flightPlanAltitude has not been s
     pilot._fms.flightPlanAltitude = previousFlightPlanAltitude;
 });
 
-ava('.climbViaSID() returns early when the given altitude is below the aircraft\'s current altitude', (t) => {
+ava('.climbViaSID() returns early when the aircraft is already above the top altitude', (t) => {
     const aircraftModel = new AircraftModel(DEPARTURE_AIRCRAFT_INIT_PROPS_MOCK);
-    aircraftModel.altitude = 5000;
+    aircraftModel.altitude = 13000;
+    const topAltitude = 5000;
     const pilot = createPilotFixture();
+    pilot._fms.departureAirportModel = airportModelFixture;
     const expectedResponse = [
         false,
         {
-            log: 'unable to comply, say again',
-            say: 'unable to comply, say again'
+            log: 'unable, we\'re already at 13000',
+            say: 'unable, we\'re already at one three thousand'
         }
     ];
 
-    const response = pilot.climbViaSid(aircraftModel, 2000);
+    const response = pilot.climbViaSid(aircraftModel, topAltitude);
+
     t.deepEqual(response, expectedResponse);
 });
 
@@ -417,11 +512,13 @@ ava('.climbViaSID() correctly configures MCP and returns correct response when n
     const aircraftModel = new AircraftModel(DEPARTURE_AIRCRAFT_INIT_PROPS_MOCK);
     aircraftModel.altitude = 0;
     const pilot = new Pilot(createFmsDepartureFixture(), createModeControllerFixture(), createNavigationLibraryFixture());
+    pilot._fms.departureAirportModel = airportModelFixture;
+    pilot._fms.flightPlanAltitude = 41000;
     const expectedResponse = [
         true,
         {
-            log: 'climb via SID and maintain 41000',
-            say: 'climb via SID and maintain flight level four one zero'
+            log: 'climb via SID and maintain 19000',
+            say: 'climb via SID and maintain flight level one niner zero'
         }
     ];
 
@@ -430,13 +527,15 @@ ava('.climbViaSID() correctly configures MCP and returns correct response when n
     t.deepEqual(response, expectedResponse);
     t.true(pilot._mcp.altitudeMode === 'VNAV');
     t.true(pilot._mcp.speedMode === 'VNAV');
-    t.true(pilot._mcp.altitude === 41000);
+    t.true(pilot._mcp.altitude === pilot._fms.departureAirportModel.maxAssignableAltitude);
 });
 
 ava('.climbViaSID() correctly configures MCP and returns correct response when an altitude is given', (t) => {
     const aircraftModel = new AircraftModel(DEPARTURE_AIRCRAFT_INIT_PROPS_MOCK);
     aircraftModel.altitude = 0;
     const pilot = new Pilot(createFmsDepartureFixture(), createModeControllerFixture(), createNavigationLibraryFixture());
+    pilot._fms.departureAirportModel = airportModelFixture;
+    pilot._fms.flightPlanAltitude = 41000;
     const expectedResponse = [
         true,
         {
@@ -463,10 +562,10 @@ ava('.conductInstrumentApproach() returns failure message when no runway is prov
 
 ava('.conductInstrumentApproach() returns failure message when assigned altitude is lower than minimum glideslope intercept altitude', (t) => {
     const expectedResult = [false, {
-        log: 'unable ILS 19L, our assigned altitude is below the minimum glideslope ' +
-        'intercept altitude, request climb to 3400',
-        say: 'unable ILS one niner left, our assigned altitude is below the minimum ' +
-        'glideslope intercept altitude, request climb to three thousand four hundred'
+        log: 'unable ILS 19L, our assigned altitude is below the minimum glideslope '
+            + 'intercept altitude, request climb to 3400',
+        say: 'unable ILS one niner left, our assigned altitude is below the minimum '
+            + 'glideslope intercept altitude, request climb to three thousand four hundred'
     }];
     const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK, createNavigationLibraryFixture());
 
@@ -508,14 +607,14 @@ ava('.conductInstrumentApproach() calls ._interceptGlidepath() with the correct 
     ));
 });
 
-ava('.conductInstrumentApproach() calls .exitHold', (t) => {
+ava('.conductInstrumentApproach() calls .cancelHoldingPattern', (t) => {
     const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK, createNavigationLibraryFixture());
-    const exitHoldSpy = sinon.spy(aircraftModel.pilot, 'exitHold');
+    const cancelHoldingPatternSpy = sinon.spy(aircraftModel.pilot, 'cancelHoldingPattern');
 
     aircraftModel.pilot._fms.setFlightPhase('HOLD');
     aircraftModel.pilot.conductInstrumentApproach(aircraftModel, approachTypeMock, runwayModelMock);
 
-    t.true(exitHoldSpy.calledWithExactly());
+    t.true(cancelHoldingPatternSpy.calledWithExactly());
 });
 
 ava('.conductInstrumentApproach() sets #hasApproachClearance to true', (t) => {
@@ -539,6 +638,65 @@ ava('.conductInstrumentApproach() returns a success message', (t) => {
     t.deepEqual(result, expectedResult);
 });
 
+ava('.crossFix() returns early when the specified fix does not exist', (t) => {
+    const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
+    const invalidFixNameMock = 'threeve';
+    const altitudeMock = 13000;
+    const expectedResult = [false, 'unable to find \'threeve\''];
+    const result = aircraftModel.pilot.crossFix(aircraftModel, invalidFixNameMock, altitudeMock);
+
+    t.deepEqual(result, expectedResult);
+});
+
+ava('.crossFix() returns early when the specified fix exists but is not on the aircraft\'s route', (t) => {
+    const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
+    const nonRouteFixNameMock = 'ogkij';
+    const altitudeMock = 13000;
+    const expectedResult = [false, 'unable, \'ogkij\' is not on our route'];
+    const result = aircraftModel.pilot.crossFix(aircraftModel, nonRouteFixNameMock, altitudeMock);
+
+    t.deepEqual(result, expectedResult);
+});
+
+ava('.crossFix() returns early when the specified altitude fails .validateNextAltitude()', (t) => {
+    const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
+    const fixNameMock = 'kepec';
+    const altitudeMock = 13000;
+    const expectedResult = [false, 'some reply from validate-next-altitude'];
+
+    sinon.stub(aircraftModel, 'validateNextAltitude').returns(expectedResult);
+
+    const result = aircraftModel.pilot.crossFix(aircraftModel, fixNameMock, altitudeMock);
+
+    t.deepEqual(result, expectedResult);
+});
+
+ava('.crossFix() correctly configures arrival aircraft\'s MCP and returns correct response when provided valid parameters', (t) => {
+    const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
+    const fixNameMock = 'kepec';
+    const altitudeMock = 13000;
+    const expectedResult = [true, {
+        log: 'cross KEPEC at 13000',
+        say: 'cross kepec at one three thousand'
+    }];
+    const result = aircraftModel.pilot.crossFix(aircraftModel, fixNameMock, altitudeMock);
+
+    t.deepEqual(result, expectedResult);
+});
+
+ava('.crossFix() correctly configures departure aircraft\'s MCP and returns correct response when provided valid parameters', (t) => {
+    const aircraftModel = new AircraftModel(DEPARTURE_AIRCRAFT_INIT_PROPS_MOCK);
+    const fixNameMock = 'cowby';
+    const altitudeMock = 13000;
+    const expectedResult = [true, {
+        log: 'cross COWBY at 13000',
+        say: 'cross cowby at one three thousand'
+    }];
+    const result = aircraftModel.pilot.crossFix(aircraftModel, fixNameMock, altitudeMock);
+
+    t.deepEqual(result, expectedResult);
+});
+
 ava('.descendViaStar() returns early when provided bottom altitude parameter is invalid', (t) => {
     const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
     const pilot = createPilotFixture();
@@ -546,7 +704,7 @@ ava('.descendViaStar() returns early when provided bottom altitude parameter is 
     pilot._mcp.setAltitudeFieldValue(initialAltitudeMock);
     pilot._mcp.setAltitudeHold();
 
-    const expectedResponse = [false, 'unable to descend to bottom altitude of threeve'];
+    const expectedResponse = [false, 'unable to maintain an altitude of threeve'];
     const response = pilot.descendViaStar(aircraftModel, invalidAltitudeMock);
 
     t.deepEqual(response, expectedResponse);
@@ -556,7 +714,7 @@ ava('.descendViaStar() returns early when provided bottom altitude parameter is 
 ava('.descendViaStar() returns early when no bottom altitude param provided and FMS has no bottom altitude', (t) => {
     const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
     const pilot = createPilotFixture();
-    const failureResponseMock = [false, 'unable to descend via STAR'];
+    const failureResponseMock = [false, 'unable, no altitude assigned'];
 
     pilot._mcp.setAltitudeFieldValue(initialAltitudeMock);
     pilot._mcp.setAltitudeHold();
@@ -573,7 +731,7 @@ ava('.descendViaStar() returns early when no bottom altitude param provided and 
 ava('.descendViaStar() returns early when no bottom altitude param provided and FMS bottom altitude is invalid', (t) => {
     const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
     const pilot = createPilotFixture();
-    const failureResponseMock = [false, 'unable to descend via STAR'];
+    const failureResponseMock = [false, 'unable, no altitude assigned'];
 
     pilot._mcp.setAltitudeFieldValue(initialAltitudeMock);
     pilot._mcp.setAltitudeHold();
@@ -592,15 +750,16 @@ ava('.descendViaStar() returns early when no bottom altitude param provided and 
 ava('.descendViaStar() returns early when the bottom altitude is above the aircraft\'s current altitude', (t) => {
     const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
     const pilot = createPilotFixture();
+    aircraftModel.altitude = 5000;
     const expectedResponse = [
         false,
         {
-            log: 'unable to comply, say again',
-            say: 'unable to comply, say again'
+            log: 'unable, we\'re already at 5000',
+            say: 'unable, we\'re already at five thousand'
         }
     ];
 
-    const response = pilot.descendViaStar(aircraftModel, 31000);
+    const response = pilot.descendViaStar(aircraftModel, 17000);
     t.deepEqual(response, expectedResponse);
 });
 
@@ -675,16 +834,22 @@ ava('.goAround() returns a success message', (t) => {
 });
 
 ava('.initiateHoldingPattern() returns error response when specified fix is not in the route', (t) => {
-    const expectedResult = [false, 'unable to hold at COWBY; it is not on our route!'];
     const pilot = createPilotFixture();
+    const expectedResult = [false, {
+        log: 'unable to hold at COWBY; it is not on our route!',
+        say: 'unable to hold at cowby; it is not on our route!'
+    }];
     const result = pilot.initiateHoldingPattern('COWBY', holdParametersMock);
 
     t.deepEqual(result, expectedResult);
 });
 
 ava('.initiateHoldingPattern() returns correct readback when hold implemented successfully', (t) => {
-    const expectedResult = [true, 'hold east of KEPEC, right turns, 1min legs'];
     const pilot = createPilotFixture();
+    const expectedResult = [true, {
+        log: 'hold east of KEPEC, right turns, 1min legs',
+        say: 'hold east of kepec, right turns, 1min legs'
+    }];
     const result = pilot.initiateHoldingPattern('KEPEC', holdParametersMock);
 
     t.deepEqual(result, expectedResult);
@@ -695,7 +860,7 @@ ava('.maintainAltitude() returns early responding that they are unable to mainta
     const nextAltitudeMock = 90000;
     const shouldExpediteMock = false;
     const shouldUseSoftCeilingMock = true;
-    const mcp = aircraftModel.mcp;
+    const { mcp } = aircraftModel;
     const expectedAltitude = mcp.altitude;
     const expectedResult = [
         false,
@@ -718,7 +883,25 @@ ava('.maintainAltitude() returns early responding that they are unable to mainta
     t.deepEqual(result, expectedResult);
 });
 
-ava('.maintainAltitude() should set mcp.altitudeMode to `HOLD` and set mcp.altitude to the correct value', (t) => {
+ava('.maintainAltitude() adds 1 additional foot to assigned altitude when assigned top altitude and "shouldUseSoftCeiling" is true', (t) => {
+    const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
+    const nextAltitudeMock = 19000;
+    const shouldExpediteMock = false;
+    const shouldUseSoftCeilingMock = true;
+
+    aircraftModel.pilot.maintainAltitude(
+        nextAltitudeMock,
+        shouldExpediteMock,
+        shouldUseSoftCeilingMock,
+        airportModelFixture,
+        aircraftModel
+    );
+
+    t.true(aircraftModel.mcp.altitudeMode === 'HOLD');
+    t.true(aircraftModel.mcp.altitude === 19001);
+});
+
+ava('.maintainAltitude() sets mcp.altitudeMode to `HOLD` and set mcp.altitude to the correct value', (t) => {
     const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
     const nextAltitudeMock = 13000;
     const shouldExpediteMock = false;
@@ -826,14 +1009,14 @@ ava('.maintainHeading() sets the #mcp with the correct modes and values', (t) =>
     t.true(aircraftModel.pilot._mcp.heading === 3.141592653589793);
 });
 
-ava('.maintainHeading() calls .exitHold()', (t) => {
+ava('.maintainHeading() calls .cancelHoldingPattern()', (t) => {
     const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
-    const exitHoldSpy = sinon.spy(aircraftModel.pilot, 'exitHold');
+    const cancelHoldingPatternSpy = sinon.spy(aircraftModel.pilot, 'cancelHoldingPattern');
 
     aircraftModel.pilot._fms.setFlightPhase('HOLD');
     aircraftModel.pilot.maintainHeading(aircraftModel, nextHeadingDegreesMock, null, false);
 
-    t.true(exitHoldSpy.calledWithExactly());
+    t.true(cancelHoldingPatternSpy.calledWithExactly());
 });
 
 ava('.maintainHeading() returns a success message when incremental is false and no direction is provided', (t) => {
@@ -876,6 +1059,38 @@ ava('.maintainHeading() returns a success message when incremental is true and d
     ];
     const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
     const result = aircraftModel.pilot.maintainHeading(aircraftModel, 42, directionMock, true);
+
+    t.deepEqual(result, expectedResult);
+});
+
+ava('.maintainHeading() returns a success message when incremental is false and direction is provided', (t) => {
+    const directionMock = 'right';
+    const isIncremental = false;
+    const expectedResult = [
+        true,
+        {
+            log: 'turn right heading 042',
+            say: 'turn right heading zero four two'
+        }
+    ];
+    const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
+    const result = aircraftModel.pilot.maintainHeading(aircraftModel, 42, directionMock, isIncremental);
+
+    t.deepEqual(result, expectedResult);
+});
+
+ava('.maintainHeading() returns a success message when incremental is false and no direction is provided', (t) => {
+    const directionMock = '';
+    const isIncremental = false;
+    const expectedResult = [
+        true,
+        {
+            log: 'fly heading 042',
+            say: 'fly heading zero four two'
+        }
+    ];
+    const aircraftModel = new AircraftModel(ARRIVAL_AIRCRAFT_INIT_PROPS_MOCK);
+    const result = aircraftModel.pilot.maintainHeading(aircraftModel, 42, directionMock, isIncremental);
 
     t.deepEqual(result, expectedResult);
 });
@@ -990,14 +1205,14 @@ ava('.proceedDirect() sets the correct #_mcp mode', (t) => {
     t.true(pilot._mcp.headingMode === 'LNAV');
 });
 
-ava('.proceedDirect() calls .exitHold()', (t) => {
+ava('.proceedDirect() calls .cancelHoldingPattern()', (t) => {
     const pilot = createPilotFixture();
-    const exitHoldSpy = sinon.spy(pilot, 'exitHold');
+    const cancelHoldingPatternSpy = sinon.spy(pilot, 'cancelHoldingPattern');
 
     pilot._fms.setFlightPhase('HOLD');
     pilot.proceedDirect(waypointNameMock);
 
-    t.true(exitHoldSpy.calledWithExactly());
+    t.true(cancelHoldingPatternSpy.calledWithExactly());
 });
 
 ava('.proceedDirect() returns success message when finished', (t) => {